ci: remove constraint on container metadata step (#11)
The step in the dev container build process which pulls metadata to generate a set of tags should not be skipped on main. The original intention was to use a separate process for tagging builds from main, but that can be merged into the metadata step so it should not be skipped.